Add 'others' section to sankey chart
