+++ src/pages/Tx/Information/Summary/index.tsx
@@ -2,7 +2,7 @@ import { ArrowRightIcon, CheckCircledIcon } from "@radix-ui/react-icons";
 import { GetTokenOutput } from "@xlabs-libs/wormscan-sdk";
 import { BlockchainIcon } from "src/components/atoms";
 import Chip from "src/components/atoms/Chip";
-import { colorStatus } from "src/conts";
+import { colorStatus } from "src/consts";
 import { TxStatus } from "src/types";
 import { formatUnits } from "src/utils/crypto";
 import "./styles.scss";
